Move xcom from /gaming to /games
