+++ .factory/automation.yml
@@ -75,7 +75,7 @@ build:
         bazel run //:typedb-extractor -- dist/typedb-all-linux
         sudo systemd-run ./dist/typedb-all-linux/typedb server
         sleep 60
-        bazel test //gaming/xcom:test --test_output=errors
+        bazel test //games/xcom:test --test_output=errors
   performance:
     test-performance-typedb-core-server:
       machine: 16-core-64-gb
@@ -165,4 +165,4 @@ build:
         bazel run //:typedb-extractor -- dist/typedb-all-linux
         sudo systemd-run ./dist/typedb-all-linux/typedb server
         sleep 60
-        bazel test //gaming/xcom:test --test_output=errors --runs_per_test=20 --jobs=1
+        bazel test //games/xcom:test --test_output=errors --runs_per_test=20 --jobs=1
