+++ README.md
@@ -30,10 +30,10 @@ Read the [README](commerce/bookstore/README.md) file for instructions. Check [th
 or the initial [dataset](commerce/bookstore/python/data) for additional information. All logic accessible in the script
 files in the [python](commerce/bookstore/python) directory.
 
-### [Gaming: XCOM Project](gaming/xcom)
+### [Games: XCOM Project](games/xcom)
 
 The XCOM 2 example contains a database of interdependent research tasks in the game XCOM 2, featuring automatic
-inference of available research based on completed tasks and available items. See [the schema](gaming/xcom/schema.tql)
+inference of available research based on completed tasks and available items. See [the schema](games/xcom/schema.tql)
 for the examples of reasoner rules inferring attributes.
 
 ### [Software: GitHub](software/github)
