+++ games/xcom/TransactionMode.java
@@ -19,14 +19,9 @@
  * under the License.
  */
 
-package com.vaticle.typedb.example.gaming.xcom;
+package com.vaticle.typedb.example.games.xcom;
 
-public class ResearchTask {
-    public ResearchTask(String name, double progressPercent) {
-        this.name = name;
-        this.progressPercent = progressPercent;
-    }
-
-    public String name;
-    public double progressPercent;
+public enum TransactionMode {
+    READ,
+    WRITE
 }
