+++ games/xcom/data/resource.csv
