+++ games/xcom/data/tech_required_event.csv
