+++ games/xcom/data/tech_required_resource.csv
