+++ games/xcom/data/tech_required_tech.csv
