+++ games/xcom/schema.tql
